requirements.txt
attrs==23.2.0
blinker==1.7.0
blis==0.7.11
catalogue==1.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cymem==2.0.8
docx2txt==0.8
Flask==3.0.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
MarkupSafe==2.1.5
murmurhash==1.0.10
nltk==3.8.1
numpy==1.24.4
pandas==2.0.3
pdfminer.six==20231228
pillow==10.2.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
preshed==3.0.9
pycparser==2.21
pycryptodome==3.20.0
pypdf==4.1.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sortedcontainers==2.4.0
spacy==2.3.5
srsly==1.0.7
thinc==7.4.6
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==0.10.1
Werkzeug==3.0.1
zipp==3.18.1
